make API: separate analysis from running wrapper
